refactor: split doctor_checks/ruby/review under 500-SLOC cap (#1195)
Partial line-cap cleanup (3 of 14 tracked files). Behaviour-preserving
splits into the repo's standard module-directory + facade pattern; all
public paths preserved via re-exports.

- doctor_checks.rs (503→322 SLOC prod): inline test module lifted to a
  sibling tests.rs.
- ruby.rs (545→379 SLOC prod): node builders → ruby/nodes.rs, call/require
  extraction → ruby/calls.rs; analyzer + walker stay in ruby/mod.rs.
- review.rs (555→227 SLOC prod): report data types → review/types.rs, diff
  parser → review/parser.rs, tests → review/tests.rs; analysis + facade in
  review/mod.rs.

Regenerated .line-cap-allowlist.tsv (14→11 grandfathered entries).
