+++ crates/trusty-analyze/src/lang/adapters/ruby/mod.rs
@@ -23,10 +23,15 @@
 //! Test: see the `tests` module — covers detection, methods, singleton
 //! methods, modules, call extraction, deduplication, and require imports.
 
-use crate::lang::call_target::build_call_target;
 use crate::lang::{LanguageAnalyzer, StaticAnalysisResult};
-use crate::types::{CodeChunk, KgEdge, KgEdgeKind, KgGraph, KgNode, KgNodeKind};
+use crate::types::{CodeChunk, KgEdge, KgEdgeKind, KgGraph, KgNodeKind};
 use tree_sitter::{Node, Parser};
+
+mod calls;
+mod nodes;
+
+use calls::{emit_require, extract_calls, is_require_call};
+use nodes::{file_node, make_method_node, make_simple_node, name_of, node_text};
 
 /// tree-sitter-ruby-backed analyzer.
 pub struct RubyAnalyzer;
@@ -94,124 +99,6 @@ impl LanguageAnalyzer for RubyAnalyzer {
 
         result
     }
-}
-
-fn file_node(file: &str) -> KgNode {
-    KgNode {
-        id: format!("ruby:File:{file}"),
-        kind: KgNodeKind::File,
-        name: file.to_string(),
-        qualified_name: file.to_string(),
-        language: "ruby".into(),
-        file: file.to_string(),
-        start_line: 0,
-        end_line: 0,
-        doc_comment: None,
-        is_public: false,
-        extra: serde_json::Value::Null,
-    }
-}
-
-fn node_text(node: Node, src: &[u8]) -> String {
-    node.utf8_text(src).unwrap_or("").to_string()
-}
-
-fn name_of(node: Node, src: &[u8]) -> Option<String> {
-    node.child_by_field_name("name").map(|n| node_text(n, src))
-}
-
-fn make_simple_node(kind: KgNodeKind, name: &str, chunk: &CodeChunk, ast: Node) -> KgNode {
-    let start = (chunk.start_line as u32).saturating_add(ast.start_position().row as u32);
-    let end = (chunk.start_line as u32).saturating_add(ast.end_position().row as u32);
-    let kind_str = format!("{kind:?}");
-    // Ruby visibility is dynamic; treat anything not starting with `_` as public.
-    let is_public = !name.starts_with('_');
-    KgNode {
-        id: format!("ruby:{kind_str}:{}:{name}", chunk.file),
-        kind,
-        name: name.to_string(),
-        qualified_name: name.to_string(),
-        language: "ruby".into(),
-        file: chunk.file.clone(),
-        start_line: start,
-        end_line: end,
-        doc_comment: None,
-        is_public,
-        extra: serde_json::Value::Null,
-    }
-}
-
-/// Build a method node with class-qualified ID/qualified_name. Mirrors the
-/// Python adapter's strategy so methods on different classes don't collide.
-///
-/// Why: Ruby method names are short and frequently shared across classes
-/// (e.g. `initialize`, `to_s`, `call`); without a class qualifier the linker
-/// would merge them.
-/// What: Returns a `KgNode` with `id = ruby:Method:file:Class:name` and
-/// `qualified_name = Class.name`. When `class_name` is empty, falls back to
-/// the bare name (top-level `def`).
-/// Test: `ruby_extracts_class_methods_with_qualified_ids`,
-/// `ruby_singleton_method_uses_class_qualified_name`.
-fn make_method_node(class_name: &str, name: &str, chunk: &CodeChunk, ast: Node) -> KgNode {
-    let start = (chunk.start_line as u32).saturating_add(ast.start_position().row as u32);
-    let end = (chunk.start_line as u32).saturating_add(ast.end_position().row as u32);
-    let qualified = if class_name.is_empty() {
-        name.to_string()
-    } else {
-        format!("{class_name}.{name}")
-    };
-    let id_suffix = if class_name.is_empty() {
-        name.to_string()
-    } else {
-        format!("{class_name}:{name}")
-    };
-    KgNode {
-        id: format!("ruby:Method:{}:{id_suffix}", chunk.file),
-        kind: KgNodeKind::Method,
-        name: name.to_string(),
-        qualified_name: qualified,
-        language: "ruby".into(),
-        file: chunk.file.clone(),
-        start_line: start,
-        end_line: end,
-        doc_comment: None,
-        // Ruby has no `_` privacy convention by default, but we mirror the
-        // other adapters so internal helpers like `_internal` are flagged.
-        is_public: !name.starts_with('_'),
-        extra: serde_json::Value::Null,
-    }
-}
-
-/// True if this `call` node is a `require` / `require_relative` invocation.
-/// We treat these as imports rather than as call edges.
-fn is_require_call(call: Node, src: &[u8]) -> bool {
-    let Some(method) = call.child_by_field_name("method") else {
-        return false;
-    };
-    if method.kind() != "identifier" {
-        return false;
-    }
-    let name = node_text(method, src);
-    name == "require" || name == "require_relative"
-}
-
-/// Names that look like declarative DSL (`attr_reader`, etc.) and shouldn't
-/// be treated as outgoing call edges.
-fn is_declarative_call(name: &str) -> bool {
-    matches!(
-        name,
-        "require"
-            | "require_relative"
-            | "attr_reader"
-            | "attr_writer"
-            | "attr_accessor"
-            | "private"
-            | "public"
-            | "protected"
-            | "include"
-            | "extend"
-            | "prepend"
-    )
 }
 
 /// Walk the Ruby AST emitting nodes/edges, keeping track of the enclosing
@@ -334,130 +221,10 @@ fn recurse(
     }
 }
 
-/// Emit one `Import` node + `Imports` edge for a `require` / `require_relative`
-/// call.
-///
-/// Why: Ruby doesn't have dedicated import syntax; requires are parsed as
-/// regular method calls. We pull them out so the dependency graph reflects
-/// file-level loading just like Python's `import`.
-/// What: Looks at the first string argument and emits one node per require.
-/// Falls back silently if the argument shape is unexpected (e.g. dynamic
-/// `require some_var`).
-/// Test: `ruby_extracts_require_imports`.
-fn emit_require(node: Node, src: &[u8], chunk: &CodeChunk, graph: &mut KgGraph, parent_id: &str) {
-    let Some(args) = node.child_by_field_name("arguments") else {
-        return;
-    };
-    let mut cursor = args.walk();
-    for child in args.children(&mut cursor) {
-        if child.kind() != "string" {
-            continue;
-        }
-        // Find the inner string_content if present, otherwise strip quotes.
-        let mut content_cursor = child.walk();
-        let mut target: Option<String> = None;
-        for inner in child.children(&mut content_cursor) {
-            if inner.kind() == "string_content" {
-                target = Some(node_text(inner, src));
-                break;
-            }
-        }
-        let target = target.unwrap_or_else(|| {
-            let raw = node_text(child, src);
-            raw.trim_matches(|c| c == '"' || c == '\'').to_string()
-        });
-        if target.is_empty() {
-            continue;
-        }
-        let n = make_simple_node(KgNodeKind::Import, &target, chunk, node);
-        let id = n.id.clone();
-        graph.nodes.push(n);
-        graph.edges.push(KgEdge {
-            from: parent_id.to_string(),
-            to: id,
-            kind: KgEdgeKind::Imports,
-            weight: 1.0,
-        });
-        // Only emit one import per require call (require takes one arg).
-        break;
-    }
-}
-
-/// Extract `call` expression nodes from a method body and produce
-/// deduplicated `Calls` edges keyed by callee name.
-///
-/// Why: Per-method outgoing call graphs are the cheapest behavioral signal we
-/// can emit; unique-per-callee deduplication with `weight = count` keeps the
-/// graph compact while preserving frequency information.
-/// What: Walks the AST subtree rooted at `body`, collects every direct `call`
-/// (skipping nested `method` / `singleton_method` / `class` / `module` /
-/// `do_block` / `block` bodies so each method only attributes its own direct
-/// calls), resolves the callee name, skips declarative DSL like `attr_reader`,
-/// counts repeats, and returns one `KgEdge` per unique callee with
-/// `weight = call_count as f32`.
-/// Test: `ruby_adapter_extracts_call_edges`,
-/// `ruby_adapter_deduplicates_repeated_calls`.
-fn extract_calls(body: Node, src: &[u8], caller_id: &str) -> Vec<KgEdge> {
-    use std::collections::HashMap;
-
-    let mut counts: HashMap<String, u32> = HashMap::new();
-
-    fn visit(node: Node, src: &[u8], counts: &mut HashMap<String, u32>) {
-        match node.kind() {
-            // Stop at nested function-like bodies so each method only
-            // attributes its own direct calls.
-            "method" | "singleton_method" | "class" | "module" | "do_block" | "block"
-            | "lambda" => {
-                return;
-            }
-            "call" => {
-                if let Some(callee) = callee_name(node, src) {
-                    if !is_declarative_call(&callee) && callee != "self" {
-                        *counts.entry(callee).or_insert(0) += 1;
-                    }
-                }
-                // Recurse into arguments so nested calls are still counted
-                // (e.g. `foo(bar())` records both foo and bar).
-            }
-            _ => {}
-        }
-        let mut cursor = node.walk();
-        for child in node.children(&mut cursor) {
-            visit(child, src, counts);
-        }
-    }
-
-    visit(body, src, &mut counts);
-
-    counts
-        .into_iter()
-        .map(|(callee, count)| KgEdge {
-            from: caller_id.to_string(),
-            to: build_call_target("ruby", "Method", &callee),
-            kind: KgEdgeKind::Calls,
-            weight: count as f32,
-        })
-        .collect()
-}
-
-/// Extract a best-effort callee name from a Ruby `call` node.
-///
-/// Why: Cross-file resolution is out of scope for the adapter (the linker
-/// merges by qualified_name later). We only need a stable string handle.
-/// What: Inspects the `method` field. Returns the bare text for `identifier`,
-/// `constant`, or `operator`; returns `None` for unsupported forms.
-/// Test: Exercised indirectly by `ruby_adapter_extracts_call_edges`.
-fn callee_name(call: Node, src: &[u8]) -> Option<String> {
-    let m = call.child_by_field_name("method")?;
-    match m.kind() {
-        "identifier" | "constant" | "operator" => Some(node_text(m, src)),
-        _ => None,
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::types::KgNode;
 
     fn make_chunk(content: &str) -> CodeChunk {
         CodeChunk {
